requirements.txt
flask
flask-jwt-extended
flask-marshmallow
marshmallow
passlib
python-dotenv
flask-cors
pytz
flask-limiter
tensorflow
numpy
pillow
opencv-python
scikit-learn
mongoengine
flask-mongoengine